Speed up loading sun
